Refactor exception handling and simplify data structures
Removed several exception classes from `Exceptions.cs`, indicating a refactor in error handling for simulation data requests and simulator connections. Simplified the `DEFINITION` enum in `SimVarRequest.cs` by removing unused data type values. Introduced new memory layout structures, `SimVarStruct` and `SimVarStringStruct`, in `SimVarRequestManager.cs` for better interoperability. Removed the `Clear()` method from `SimVarRequestManager`, suggesting a change in request management. Also, eliminated redundant structure definitions in `SimVarResults.cs`.
